+++ R/get_predicted_bayesian.R
@@ -54,6 +54,7 @@ get_predicted.stanreg <- function(x,
   fun_args <- list(x,
     newdata = my_args$data,
     re.form = my_args$re.form,
+    dpar = my_args$distributional_parameter,
     draws = iterations
   )
 
@@ -86,7 +87,12 @@ get_predicted.stanreg <- function(x,
   }
 
   # Get predictions (summarize)
-  predictions <- .get_predicted_centrality_from_draws(x, iter = draws, ...)
+  predictions <- .get_predicted_centrality_from_draws(
+    x,
+    iter = draws,
+    datagrid = my_args$data,
+    ...
+  )
 
   # Output
   ci_data <- get_predicted_ci(